refactor: add context types file
