+++ src/context/context.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from "@nestjs/common";
 import { getClaims, getSub } from "../auth/auth.utils";
 import { RefreshTokensService } from "../auth/refresh-tokens/refresh-tokens.service";
 import { ShieldService } from "../auth/shield/shield.service";
-import { Context, ContextServices } from "../common/common.types";
+import { Context, ContextServices } from "./context.types";
 import { DataloaderService } from "../dataloader/dataloader.service";
 import { EventsService } from "../events/events.service";
 import { GroupMemberRequestsService } from "../groups/group-member-requests/group-member-requests.service";
